More stable app.configs
